Major Refactor now using seroost.
This version is not functional but demonstrates to expected
data flow.

Now using seroost library ( model and lexer ) to precompte
the indexer.

A memoized signal is created from a users search_query.
This memo is a ranked list of images.

The code is buggy .. images are not displayed just a
list of filepath is returned.

Additionally using leptos 0.5.0 beta