Two fixes which breaks the search update looping I was having problems with.
A) Refactored the chain to derived signals :-

from the root_directory ( a signal ).
Which generates the index ( a derived signal ).

The index is queried to genrate a final derrived signal
The list of image urls to display.

Index now support PartialEq

-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct Index {
  ..
 }

B) the form input is now a controlled input...