Restored image description.
As the images are extracted on the server, the image descriptions are bundled
up in the search results (SRElem)

Not many images have a description.
To demonstrate the data flow seach for "chinook"

Tailwindcss:

Empty "ImageGallery" now looks like a bar that extends to the full width of the page.