Restored the SideBar
Now the metadata is explicity computed on the server.

Solution :-

Added new create_action/resource/DerivedSignal path

When the metadata button is clicked the new action is
dispatched and the sidebar opens.

Previously the SideBar was up one level in the ImageGallery.
But this is prefered as create_action at the same level as the button.