Restructured filelist
server function are in the file_lister/mod.rs
component in the lister.rs

Settings bar: For the first time when opened, a list of directories
is seen.

Sanitization functions which are going to be using
more than once have been moved to methods on global state.

Global STATE was public, now pub(crate)