Settings sidebar: add_list_url() and get_list_url() first attmept at dispatch.
Directory items are now a input element (readonly) so that value can be extracted.

Developed a complex code snippet for extacting the input.value() element dispatch a
url value based on the text.

This snippet need to be used in many places.
And on a higher level element like the <ol>